Leetcode 437 Path SUm III

+++ 0437_Path Sum III/PathSumIII.java
@@ -1,14 +1,14 @@
 class Solution {
     private int ans = 0;
     public int pathSum(TreeNode root, int targetSum) {
-        traverse(root, new ArrayList<Integer>(), targetSum);
+        traverse(root, new ArrayList<Long>(), targetSum);
         return ans;
     }
     
-    private void traverse(TreeNode node, List<Integer> list, int targetSum) {
+    private void traverse(TreeNode node, List<Long> list, int targetSum) {
         if(node == null) return;
-        list.add(node.val);
-        int sum = 0;
+        list.add((long)node.val);
+        long sum = 0;
         for(int i=list.size()-1;i>=0;i--) {
             sum += list.get(i);
             if(sum == targetSum) ans++;
